+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [1.0.4] - 2021-02-23
 
+### Fixed
+
+-   Auto length for float32/float64 array
+
 ### Changed
 
 -   Changed `order` parameter to `endianess` in `ByteSet`
@@ -9,8 +13,9 @@
 
 ### Added
 
--   Added `ByteArray` class. Is the same as `ByteSet` but has dynamic size. It's good for
-    write-only operations.
+-   Added `ByteArray` class. Is the same as `ByteSet` but has dynamic size. It's
+    good for write-only operations.
+-   Added both BE and LE endianess for read and write in `ByteSet`
 
 ## [1.0.3] - 2021-02-18
 
@@ -20,11 +25,14 @@
 
 ### Added
 
--   Added `print` function that logs into console byteset and show current location
+-   Added `print` function that logs into console byteset and show current
+    location
 -   Added `isEnd` property to `ByteSet`
 -   Added more Big Endian in functions
 -   Added `uint24` write and read for colors
--   Added `each` function to go throuch each buffer element. You can set number type for example `uint8` for each byte or `uint16` for each short. By deafult it goes until position is reach end of the buffer.
+-   Added `each` function to go throuch each buffer element. You can set number
+    type for example `uint8` for each byte or `uint16` for each short. By deafult
+    it goes until position is reach end of the buffer.
 -   Added `limit` parameter to `each` function
 -   Added read `int32Array` and `uint32Array`
 -   Added `slice` for `ByteSet`